problem 57 solved

+++ problems/problem_15.hpp
@@ -1,2 +1,2 @@
-int pgcd(int a, int b);
+long long pgcd(long long a, long long b);
 int problem_15();
